[ncl][bare-expo] Prepare screens for expo-router migration
Behavior-preserving prep before migrating bare-expo, native-component-list,
and test-suite from react-navigation to expo-router (SDK 56):

- Extract the api/component screen lists into pure registry modules
  (apiScreens.ts, componentScreens.ts) with id-based lookups (screenRegistry.ts),
  so file-based routes can resolve screens by deep-link id later.
- Make route params serializable: MediaLibrary passes assetId/albumId instead
  of asset/album objects and an onGoBack callback (replaced by refresh on
  focus), Calendars passes calendarId, SVG passes the example key.
- Flatten inner stack navigators: SearchScreen exports a plain screen (header
  options move to the host stack), the react-native-screens demo registers its
  sub-screens as regular routes instead of embedding navigators.
- Drop the opt-in navigation state persistence (lock toggle + AsyncStorage
  restore) since expo-router owns navigation state.
