+++ apps/native-component-list/src/navigation/ExpoApisStackNavigator.tsx
@@ -1,544 +1,17 @@
 import { BottomTabNavigationProp } from '@react-navigation/bottom-tabs';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { useTheme } from 'ThemeProvider';
-import { isRunningInExpoGo } from 'expo';
 import * as React from 'react';
 
 import { TabBackground } from '../components/TabBackground';
 import TabIcon from '../components/TabIcon';
 import getStackNavWithConfig from '../navigation/StackConfig';
-import { AudioScreens } from '../screens/Audio/AudioScreen';
-import { BlobScreens } from '../screens/Blob/BlobScreen';
-import { CalendarNextScreens } from '../screens/Calendar@Next/CalendarNextScreens';
-import { CalendarsScreens } from '../screens/CalendarsScreen';
-import { apiScreensToListElements } from '../screens/ComponentListScreen';
-import { ContactsScreens } from '../screens/Contacts/ContactsScreen';
-import { ContactsNextScreens } from '../screens/Contacts@Next/ContactsNextScreen';
-import { CryptoScreens } from '../screens/Crypto/CryptoScreen';
 import ExpoApis from '../screens/ExpoApisScreen';
-import { MediaLibraryScreens } from '../screens/MediaLibrary@Next/MediaLibraryScreens';
-import { ModulesCoreScreens } from '../screens/ModulesCore/ModulesCoreScreen';
-import { WorkletsScreens } from '../screens/Worklets/WorkletsScreen';
-import { type ScreenConfig } from '../types/ScreenConfig';
-import { optionalRequire } from './routeBuilder';
+import { ScreensList, Screens, screenApiItems } from './apiScreens';
+
+export { ScreensList, Screens, screenApiItems };
 
 const Stack = createNativeStackNavigator();
-
-export const ScreensList: ScreenConfig[] = [
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/CameraPermissions/CameraPermissionsScreen'));
-    },
-    name: 'CameraPermissions',
-    options: { title: 'Camera Permissions' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ModulesCore/ModulesCoreScreen'));
-    },
-    name: 'ModulesCore',
-    options: { title: 'Expo Modules Core' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/MediaLibrary@Next/MediaLibraryScreens'));
-    },
-    name: 'MediaLibrary@Next',
-    options: { title: 'MediaLibrary@Next' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/StatusBarScreen'));
-    },
-    name: 'StatusBar',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/AlertScreen'));
-    },
-    name: 'Alert',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Clipboard/ClipboardScreen'));
-    },
-    name: 'Clipboard',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/CellularScreen'));
-    },
-    name: 'Cellular (device-only)',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/AccelerometerScreen'));
-    },
-    name: 'Accelerometer',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ActionSheetScreen'));
-    },
-    name: 'ActionSheet',
-    options: { title: 'Action Sheet' },
-  },
-  ...(isRunningInExpoGo()
-    ? []
-    : [
-        {
-          getComponent() {
-            return optionalRequire(() => require('../screens/AgeRangeScreen'));
-          },
-          name: 'AgeRange',
-          options: { title: 'Age Range' },
-        },
-      ]),
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/AppearanceScreen'));
-    },
-    name: 'Appearance',
-  },
-  // TODO: fix this, erroring in release in Expo Go
-  // {
-  //   getComponent() {
-  //     return optionalRequire(() => require('../screens/AppIntegrity/AppIntegrityScreen'));
-  //   },
-  //   name: 'AppIntegrity',
-  // },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/AppleAuthenticationScreen'));
-    },
-    name: 'AppleAuthentication',
-    options: { title: 'Apple Authentication' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/AppMetricsScreen'));
-    },
-    name: 'AppMetrics',
-    options: { title: 'App Metrics' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ExpoObserveScreen'));
-    },
-    name: 'ExpoObserve',
-    options: { headerShown: false, title: 'Expo Observe' },
-    route: 'expo-observe',
-    linking: {
-      path: 'expo-observe',
-      screens: {
-        index: '',
-        filteredParams: 'filtered/:userId/:accountId',
-      },
-    },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/AsyncStorageScreen'));
-    },
-    name: 'AsyncStorage',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Audio/AudioScreen'));
-    },
-    name: 'Audio',
-  },
-  // TODO: fix this, erroring in release in Expo Go
-  // {
-  //   getComponent() {
-  //     return optionalRequire(() => require('../screens/AuthSession/AuthSessionScreen'));
-  //   },
-  //   name: 'AuthSession',
-  // },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Location/BackgroundLocationMapScreen'));
-    },
-    name: 'BackgroundLocation',
-    options: { title: 'Background location' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/BackgroundFetchScreen'));
-    },
-    name: 'BackgroundFetch',
-    options: { title: 'Background Fetch' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/BackgroundTaskScreen'));
-    },
-    name: 'BackgroundTask',
-    options: { title: 'Background Task' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/BatteryScreen'));
-    },
-    name: 'Battery',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/inlineModules/InlineModulesScreen'));
-    },
-    name: 'InlineModules',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Blob/BlobScreen'));
-    },
-    name: 'Blob',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/BrightnessScreen'));
-    },
-    name: 'Brightness (device-only)',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/DeviceScreen'));
-    },
-    name: 'Device',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/DocumentPickerScreen'));
-    },
-    name: 'DocumentPicker',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/LocalizationScreen'));
-    },
-    name: 'Localization',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/FileSystemScreen'));
-    },
-    name: 'FileSystem',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/FileSystemLegacyScreen'));
-    },
-    name: 'FileSystem@legacy',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/FontScreen'));
-    },
-    name: 'Font',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/HapticsScreen'));
-    },
-    name: 'Haptics',
-    options: { title: 'Haptics Feedback' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/CalendarsScreen'));
-    },
-    name: 'Calendars',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Calendar@Next/CalendarNextScreens'));
-    },
-    name: 'Calendars@next',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ConstantsScreen'));
-    },
-    name: 'Constants',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Contacts/ContactsScreen'));
-    },
-    name: 'Contacts',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Contacts@Next/ContactsNextScreen'));
-    },
-    name: 'Contacts@Next',
-    options: { title: 'Contacts@Next' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ErrorScreen'));
-    },
-    name: 'Errors',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ImageManipulatorScreen'));
-    },
-    name: 'ImageManipulator',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ImageManipulatorScreenLegacy'));
-    },
-    name: 'ImageManipulator (legacy)',
-    route: 'image-manipulator-legacy',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ImagePicker/ImagePickerScreen'));
-    },
-    name: 'ImagePicker',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/IntentLauncherScreen'));
-    },
-    name: 'IntentLauncher',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/KeepAwakeScreen'));
-    },
-    name: 'KeepAwake',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/LinkingScreen'));
-    },
-    name: 'Linking',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/MailComposerScreen'));
-    },
-    name: 'MailComposer',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/MediaLibrary/MediaLibraryScreen'));
-    },
-    name: 'MediaLibrary',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/MediaLibrary/MediaAlbumsScreen'));
-    },
-    name: 'MediaAlbums',
-    options: { title: 'MediaLibrary Albums' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/MediaLibrary/MediaDetailsScreen'));
-    },
-    name: 'MediaDetails',
-    options: { title: 'MediaLibrary Asset' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/NetInfoScreen'));
-    },
-    name: 'NetInfo',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/NetworkScreen'));
-    },
-    name: 'Network',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Crypto/CryptoScreen'));
-    },
-    name: 'Crypto',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/NotificationScreen'));
-    },
-    name: 'Notification',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/LocalAuthenticationScreen'));
-    },
-    name: 'LocalAuthentication',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Location/LocationScreen'));
-    },
-    name: 'Location',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Location/GeocodingScreen'));
-    },
-    name: 'Geocoding',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Location/GeofencingScreen'));
-    },
-    name: 'Geofencing',
-    options: { title: 'Geofencing Map' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/PedometerScreen'));
-    },
-    name: 'Pedometer',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/PrintScreen'));
-    },
-    name: 'Print',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/RemindersScreen'));
-    },
-    name: 'Reminders',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/SafeAreaContextScreen'));
-    },
-    name: 'SafeAreaContext',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ScreenOrientationScreen'));
-    },
-    name: 'ScreenOrientation',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/SecureStoreScreen'));
-    },
-    name: 'SecureStore',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ScreenCaptureScreen'));
-    },
-    name: 'ScreenCapture',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/SensorScreen'));
-    },
-    name: 'Sensor',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/SharingScreen'));
-    },
-    name: 'Sharing',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/NavigationBarScreen'));
-    },
-    name: 'NavigationBar',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/SystemUIScreen'));
-    },
-    name: 'SystemUI',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/SMSScreen'));
-    },
-    name: 'SMS',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/StoreReview'));
-    },
-    name: 'StoreReview',
-    options: { title: 'Store Review' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/TaskManagerScreen'));
-    },
-    name: 'TaskManager',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/TextToSpeechScreen'));
-    },
-    name: 'TextToSpeech',
-    options: { title: 'Speech' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/TrackingTransparencyScreen'));
-    },
-    name: 'TrackingTransparency',
-    options: { title: 'TrackingTransparency' },
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/UpdatesScreen'));
-    },
-    name: 'Updates Reload Screen',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/Worklets/WorkletsScreen'));
-    },
-    name: 'Worklets integration',
-    route: 'worklets',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/WebBrowser/WebBrowserScreen'));
-    },
-    name: 'WebBrowser',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/VideoThumbnailsScreen'));
-    },
-    name: 'Video Thumbnails',
-  },
-  {
-    getComponent() {
-      return optionalRequire(() => require('../screens/ViewShotScreen'));
-    },
-    name: 'ViewShot',
-  },
-];
-
-export const Screens: ScreenConfig[] = [
-  ...ScreensList,
-  ...ModulesCoreScreens,
-  ...MediaLibraryScreens,
-  ...AudioScreens,
-  ...BlobScreens,
-  ...ContactsScreens,
-  ...ContactsNextScreens,
-  ...CalendarsScreens,
-  ...CalendarNextScreens,
-  ...CryptoScreens,
-  ...WorkletsScreens,
-];
-
-export const screenApiItems = apiScreensToListElements(ScreensList);
 
 function ExpoApisStackNavigator(props: { navigation: BottomTabNavigationProp<any> }) {
   const { theme } = useTheme();
