+++ apps/native-component-list/src/screens/RemindersScreen.tsx
@@ -29,10 +29,12 @@ const ReminderRow: React.FunctionComponent<RowProps> = ({
 
 interface State {
   reminders: Calendar.Reminder[];
+  calendar: Calendar.Calendar | null;
 }
 
+// Route params must stay serializable, so this screen receives the id and refetches the calendar.
 type Links = {
-  Reminders: { calendar: Calendar.Calendar };
+  Reminders: { calendarId: string };
 };
 
 type Props = StackScreenProps<Links, 'Reminders'>;
@@ -44,14 +46,21 @@ export default class RemindersScreen extends React.Component<Props, State> {
 
   readonly state: State = {
     reminders: [],
+    calendar: null,
   };
 
   componentDidMount() {
-    const { params } = this.props.route;
-    if (params) {
-      this._findReminders(params.calendar.id!);
+    const calendarId = this.props.route.params?.calendarId;
+    if (calendarId) {
+      this._loadCalendar(calendarId);
+      this._findReminders(calendarId);
     }
   }
+
+  _loadCalendar = async (calendarId: string) => {
+    const calendars = await Calendar.getCalendarsAsync(Calendar.EntityTypes.REMINDER);
+    this.setState({ calendar: calendars.find(({ id }) => id === calendarId) ?? null });
+  };
 
   _findReminders = async (id: string) => {
     const reminders = await Calendar.getRemindersAsync([id], null, new Date(), new Date());
@@ -59,7 +68,10 @@ export default class RemindersScreen extends React.Component<Props, State> {
   };
 
   _addReminder = async () => {
-    const { calendar } = this.props.route.params!;
+    const { calendar } = this.state;
+    if (!calendar) {
+      return;
+    }
     if (!calendar.allowsModifications) {
       Alert.alert('This calendar does not allow modifications');
       return;
@@ -92,7 +104,10 @@ export default class RemindersScreen extends React.Component<Props, State> {
   };
 
   _updateReminder = async (reminder: Calendar.Reminder) => {
-    const { calendar } = this.props.route.params!;
+    const { calendar } = this.state;
+    if (!calendar) {
+      return;
+    }
     if (!calendar.allowsModifications) {
       Alert.alert('This calendar does not allow modifications');
       return;
@@ -112,17 +127,17 @@ export default class RemindersScreen extends React.Component<Props, State> {
 
   _deleteReminder = async (reminderId: string) => {
     try {
-      const { calendar } = this.props.route.params!;
+      const { calendarId } = this.props.route.params!;
       await Calendar.deleteReminderAsync(reminderId);
       Alert.alert('Reminder deleted successfully');
-      this._findReminders(calendar.id!);
+      this._findReminders(calendarId);
     } catch (e: any) {
       Alert.alert('Reminder not deleted successfully', e.message);
     }
   };
 
   render() {
-    if (!this.props.route.params?.calendar) {
+    if (!this.props.route.params?.calendarId) {
       return <BodyText>Access this screen from the "Calendars" screen.</BodyText>;
     }
     if (this.state.reminders.length) {
